Move 'PieceSquare' and 'DirtyPiece' definitions to types.h
